+++ src/framework/piral-base/src/index.ts
@@ -1,5 +1,6 @@
 export * from './create';
 export * from './load';
+export * from './loader';
 export * from './strategies';
 export * from './types';
 export * from './utils';